move rest of the common methods to BaseTest.
